agrego header para Evento y arreglo uso de string en los otros headers

+++ include/datatypes/DTExp.h
@@ -5,7 +5,7 @@
 
 class DTExp: public DTDistribucion {
     private:
-        unsigned float lambda;
+        float lambda;
     public:
         DTExp(float);
         float getLambda() const;
